refactor(core): standards loaders move to data/fs, core keeps one documented read
Per the design call: the standards loading functions leave the core layer.
data/fs/standards_loader.py now owns load_dimension, load_asvs_l1,
load_cisq, load_compiled_refs_multi, load_compiled_requirements(_multi)
and build_req_refs_lookup (which was a passthrough in core/evidence).
core/standards/loader.py is deleted; core/standards/refs.py keeps the
pure extraction functions.

ONE read stays in core, deliberately and documented in place:
core/evidence/_refs.enrich_judgment resolves requirement refs while
PARSING judgments, and compiled_dir reaches it from 15 external call
sites on the evidence path that feeds scoring. Threading a loader
through all 15 would be ceremony at every site (each passes the same
function) and the real fix is to lift enrichment out of the parse path
— a separate workstream with its own behavior-preservation burden.
So core/standards/refs.py retains _load_compiled_data +
load_compiled_refs, with a comment stating exactly why and what would
remove them.

Net: core's filesystem surface goes from 8 loader functions to 1.
