+++ src/quodeq/core/standards/__init__.py
@@ -1,18 +1,19 @@
-"""Standards loading and reference extraction utilities."""
+"""Standards reference extraction utilities (pure).
 
-from quodeq.core.standards.loader import (
-    load_asvs_l1,
-    load_cisq,
-    load_dimension,
+Loading standards from disk lives in ``data/fs/standards_loader.py``: the
+core layer holds only the parsing/extraction logic.
+"""
+
+from quodeq.core.standards.refs import (
+    extract_refs,
+    extract_requirements,
+    load_compiled_refs,
+    ref_label,
 )
-from quodeq.core.standards.refs import extract_refs, extract_requirements, load_compiled_refs, ref_label
 
 __all__ = [
     "extract_refs",
     "extract_requirements",
-    "load_asvs_l1",
-    "load_cisq",
     "load_compiled_refs",
-    "load_dimension",
     "ref_label",
 ]
