+++ src/quodeq/core/standards/refs.py
@@ -90,7 +90,13 @@ def extract_requirements(data: dict, overrides: dict[str, dict] | None = None) -
 
 
 # ---------------------------------------------------------------------------
-# I/O convenience wrappers (kept for backward-compatible call-sites)
+# The one I/O function core still owns.
+#
+# ``core/evidence/_refs.enrich_judgment`` resolves requirement refs while
+# parsing judgments, and ``compiled_dir`` is threaded to it from 15 call
+# sites. Until enrichment moves out of the parse path (its own workstream),
+# core needs this read. Every OTHER standards loader lives in
+# ``data/fs/standards_loader.py``.
 # ---------------------------------------------------------------------------
 
 def _load_compiled_data(
@@ -100,8 +106,6 @@ def _load_compiled_data(
     """Load raw compiled standards JSON from *compiled_dir*. Returns None on error.
 
     Falls back to *evaluators_dir* for custom evaluators when provided.
-    This is an I/O adapter — callers that already have the data should use
-    :func:`extract_refs` or :func:`extract_requirements` directly.
     """
     if not dimension:
         return None
@@ -113,7 +117,6 @@ def _load_compiled_data(
             except (OSError, ValueError, UnicodeDecodeError) as exc:
                 _logger.warning("Failed to load compiled standards for %s: %s", dimension, exc)
                 return None
-    # Fallback: custom evaluators directory (caller must supply it)
     if evaluators_dir:
         evaluators_path = evaluators_dir / f"{dimension}.json"
         if evaluators_path.is_file():
@@ -128,60 +131,8 @@ def load_compiled_refs(
     compiled_dir: str | Path | None, dimension: str | None,
     evaluators_dir: Path | None = None,
 ) -> dict[str, list[dict]]:
-    """Load {req_id: [{label, url, ...}, ...]} from compiled standards on disk.
-
-    Backward-compat convenience wrapper that handles file I/O then delegates
-    to the pure :func:`extract_refs`.  Prefer ``extract_refs`` when data is
-    already loaded.
-    """
+    """Load ``{req_id: [{label, url, ...}, ...]}`` from compiled standards on disk."""
     data = _load_compiled_data(compiled_dir, dimension, evaluators_dir=evaluators_dir)
     if not data:
         return {}
     return extract_refs(data)
-
-
-def load_compiled_refs_multi(
-    compiled_dir: str | Path | None, dimensions: list[str],
-    evaluators_dir: Path | None = None,
-) -> dict[str, list[dict]]:
-    """Load refs for multiple dimensions, merging into a single lookup."""
-    merged: dict[str, list[dict]] = {}
-    for dim in dimensions:
-        merged.update(load_compiled_refs(compiled_dir, dim, evaluators_dir=evaluators_dir))
-    return merged
-
-
-def load_compiled_requirements_multi(
-    compiled_dir: str | Path | None, dimensions: list[str],
-    evaluators_dir: Path | None = None,
-    overrides: dict[str, dict] | None = None,
-) -> dict[str, dict]:
-    """Load requirements for multiple dimensions, merging into a single lookup."""
-    merged: dict[str, dict] = {}
-    for dim in dimensions:
-        merged.update(load_compiled_requirements(
-            compiled_dir, dim,
-            evaluators_dir=evaluators_dir,
-            overrides=overrides,
-        ))
-    return merged
-
-
-def load_compiled_requirements(
-    compiled_dir: str | Path | None, dimension: str | None,
-    evaluators_dir: Path | None = None,
-    overrides: dict[str, dict] | None = None,
-) -> dict[str, dict]:
-    """Load {req_id: {principle, text}} from compiled standards on disk.
-
-    When *overrides* is supplied, requirement text placeholders are resolved
-    using the per-requirement override values.
-
-    Backward-compat convenience wrapper that handles file I/O then delegates
-    to the pure :func:`extract_requirements`.  Used by the MCP server to
-    auto-fill principle name and requirement text from the requirement ID.
-    """
-    data = _load_compiled_data(compiled_dir, dimension, evaluators_dir=evaluators_dir)
-    if not data:
-        return {}
-    return extract_requirements(data, overrides=overrides)
