+++ src/quodeq/services/_violations_jsonl.py
@@ -8,7 +8,7 @@ from typing import Iterable
 
 from quodeq.core.types import Finding, ViolationResponse
 from quodeq.core.evidence._req_mapping import PrincipleResolver, build_principle_resolver
-from quodeq.core.evidence.parser import build_req_refs_lookup
+from quodeq.data.fs.standards_loader import build_req_refs_lookup
 from quodeq.data.fs.stream_files import count_files_in_stream
 from quodeq.services.violation_context import ViolationContext
 from quodeq.services.suppression import SuppressionMatcher, load_req_to_principle
