+++ tests/analysis/mcp/test_mcp_findings.py
@@ -10,7 +10,7 @@ from unittest.mock import patch
 import pytest
 
 from quodeq.analysis.mcp import findings_server as mcp_findings
-from quodeq.core.standards.refs import load_compiled_refs
+from quodeq.data.fs.standards_loader import load_compiled_refs
 
 from tests._evidence_helpers import _make_request, _run_server
